update to ssd code
