+++ examples/pytorch/vision/Face_Detection/data/widerface.py
@@ -1,22 +1,13 @@
-## This code is built on https://github.com/yxlijun/S3FD.pytorch
-#-*- coding:utf-8 -*-
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
 import torch
-from PIL import Image, ImageDraw
 import torch.utils.data as data
-import numpy as np
-import random
+from PIL import Image, ImageDraw
 import sys; sys.path.append('../')
 from utils.augmentations import preprocess
+import numpy as np
+
 
 
 class WIDERDetection(data.Dataset):
-    """docstring for WIDERDetection"""
-
     def __init__(self, list_file, mode='train', mono_mode=False):
         super(WIDERDetection, self).__init__()
         self.mode = mode
@@ -61,8 +52,7 @@ class WIDERDetection(data.Dataset):
         while True:
             image_path = self.fnames[index]
             img = Image.open(image_path)
-            if img.mode == 'L':
-                img = img.convert('RGB')
+            img = img.convert('RGB')
 
             im_width, im_height = img.size
             boxes = self.annotransform(
@@ -71,28 +61,9 @@ class WIDERDetection(data.Dataset):
             bbox_labels = np.hstack((label[:, np.newaxis], boxes)).tolist()
             img, sample_labels = preprocess(
                 img, bbox_labels, self.mode, image_path)
-            sample_labels = np.array(sample_labels)
-            if len(sample_labels) > 0:
-                target = np.hstack(
-                    (sample_labels[:, 1:], sample_labels[:, 0][:, np.newaxis]))
-
-                assert (target[:, 2] > target[:, 0]).any()
-                assert (target[:, 3] > target[:, 1]).any()
-                break 
-            else:
-                index = random.randrange(0, self.num_samples)
-
-        
-        #img = Image.fromarray(img)
-        '''
-        draw = ImageDraw.Draw(img)
-        w,h = img.size
-        for bbox in sample_labels:
-            bbox = (bbox[1:] * np.array([w, h, w, h])).tolist()
-
-            draw.rectangle(bbox,outline='red')
-        img.save('image.jpg')
-        '''
+            sample_labels = np.array(sample_labels)            
+            target = np.hstack(
+                (sample_labels[:, 1:], sample_labels[:, 0][:, np.newaxis]))
 
         if self.mono_mode==True:
             im = 0.299 * img[0] + 0.587 * img[1] + 0.114 * img[2]
@@ -107,31 +78,3 @@ class WIDERDetection(data.Dataset):
         boxes[:, 2] /= im_width
         boxes[:, 3] /= im_height
         return boxes
-
-
-def detection_collate(batch):
-    """Custom collate fn for dealing with batches of images that have a different
-    number of associated object annotations (bounding boxes).
-
-    Arguments:
-        batch: (tuple) A tuple of tensor images and lists of annotations
-
-    Return:
-        A tuple containing:
-            1) (tensor) batch of images stacked on their 0 dim
-            2) (list of tensors) annotations for a given image are stacked on
-                                 0 dim
-    """
-    targets = []
-    imgs = []
-    for sample in batch:
-        imgs.append(sample[0])
-        targets.append(torch.FloatTensor(sample[1]))
-    return torch.stack(imgs, 0), targets
-
-
-if __name__ == '__main__':
-    from config_real import cfg
-    dataset = WIDERDetection(cfg.FACE.TRAIN_FILE)
-    #for i in range(len(dataset)):
-    dataset.pull_item(14)
