+++ examples/pytorch/vision/Face_Detection/layers/__init__.py
@@ -1,9 +1,2 @@
-#-*- coding:utf-8 -*-
-
-from __future__ import division
-from __future__ import absolute_import
-from __future__ import print_function
-
-
 from .functions import *
 from .modules import *
