+++ examples/pytorch/vision/Face_Detection/layers/functions/prior_box.py
@@ -1,11 +1,3 @@
-## This code is built on https://github.com/yxlijun/S3FD.pytorch
-
-#-*- coding:utf-8 -*-
-
-from __future__ import division
-from __future__ import absolute_import
-from __future__ import print_function
-
 import torch
 from itertools import product as product
 import math
@@ -54,60 +46,3 @@ class PriorBox(object):
         if self.clip:
             output.clamp_(max=1, min=0)
         return output
-
-
-class PriorBoxReal(object):
-    """Compute priorbox coordinates in center-offset form for each source
-    feature map.
-    """
-
-    def __init__(self, input_size, feature_maps,cfg):
-        super(PriorBoxReal, self).__init__()
-        self.imh = input_size[0]
-        self.imw = input_size[1]
-
-        # number of priors for feature map location (either 4 or 6)
-        self.variance = cfg.VARIANCE or [0.1]
-        #self.feature_maps = cfg.FEATURE_MAPS
-        self.min_sizes = cfg.ANCHOR_SIZES
-        self.steps = cfg.STEPS
-        self.clip = cfg.CLIP
-        for v in self.variance:
-            if v <= 0:
-                raise ValueError('Variances must be greater than 0')
-        self.feature_maps = feature_maps
-
-
-    def forward(self):
-        mean = []
-        # import pdb; pdb.set_trace()
-        for k in range(len(self.feature_maps)):
-            feath = self.feature_maps[k][0]
-            featw = self.feature_maps[k][1]
-            for i, j in product(range(feath), range(featw)):
-                f_kw = self.imw / self.steps[k]
-                f_kh = self.imh / self.steps[k]
-
-                cx = (j + 0.5) / f_kw
-                cy = (i + 0.5) / f_kh
-
-                s_kw_1 = self.min_sizes[k] / self.imw
-                s_kh_1 = self.min_sizes[k] / self.imh
-
-                # s_kw_2 = self.min_sizes[2*k+1] / self.imw
-                # s_kh_2 = self.min_sizes[2*k+1] / self.imh
-
-                mean += [cx, cy, s_kw_1, s_kh_1]
-                # mean += [cx, cy, s_kw_2, s_kh_2]
-
-        output = torch.Tensor(mean).view(-1, 4)
-        if self.clip:
-            output.clamp_(max=1, min=0)
-        return output
-
-
-if __name__ == '__main__':
-    from data.config import cfg
-    p = PriorBox([640, 640], cfg)
-    out = p.forward()
-    print(out.size())
