+++ examples/pytorch/vision/Face_Detection/layers/modules/__init__.py
@@ -1,10 +1,3 @@
-#-*- coding:utf-8 -*-
-
-from __future__ import division
-from __future__ import absolute_import
-from __future__ import print_function
-
-
 from .l2norm import L2Norm
 from .multibox_loss import MultiBoxLoss, MultiBoxLossFocal
