+++ examples/pytorch/vision/Face_Detection/layers/modules/l2norm.py
@@ -1,18 +1,6 @@
-## This code is built on https://github.com/yxlijun/S3FD.pytorch
-
-#-*- coding:utf-8 -*-
-
-from __future__ import division
-from __future__ import absolute_import
-from __future__ import print_function
-
-
 import torch
 import torch.nn as nn
 import torch.nn.init as init
-# from torch.autograd import Function
-# from torch.autograd import Variable
-
 
 
 class L2Norm(nn.Module):
