+++ examples/pytorch/vision/Face_Detection/layers/modules/multibox_loss.py
@@ -1,11 +1,3 @@
-## This code is built on https://github.com/yxlijun/S3FD.pytorch
-
-#-*- coding:utf-8 -*-
-
-from __future__ import division
-from __future__ import absolute_import
-from __future__ import print_function
-
 import math
 import torch
 import torch.nn as nn
@@ -14,39 +6,6 @@ from torch.autograd import Variable
 
 
 from ..bbox_utils import match, log_sum_exp, match_ssd
-
-
-
-class FocalLoss(nn.Module):
-    def __init__(self, gamma=0, alpha=None, size_average=True):
-        super(FocalLoss, self).__init__()
-        self.gamma = gamma
-        self.alpha = alpha
-        if isinstance(alpha,(float,int)): self.alpha = torch.Tensor([alpha,1-alpha])
-        if isinstance(alpha,list): self.alpha = torch.Tensor(alpha)
-        self.size_average = size_average
-
-    def forward(self, input, target):
-        if input.dim()>2:
-            input = input.view(input.size(0),input.size(1),-1)  # N,C,H,W => N,C,H*W
-            input = input.transpose(1,2)    # N,C,H*W => N,H*W,C
-            input = input.contiguous().view(-1,input.size(2))   # N,H*W,C => N*H*W,C
-        target = target.view(-1,1)
-
-        logpt = F.log_softmax(input)
-        logpt = logpt.gather(1,target)
-        logpt = logpt.view(-1)
-        pt = Variable(logpt.data.exp())
-
-        if self.alpha is not None:
-            if self.alpha.type()!=input.data.type():
-                self.alpha = self.alpha.type_as(input.data)
-            at = self.alpha.gather(0,target.data.view(-1))
-            logpt = logpt * Variable(at)
-
-        loss = -1 * (1-pt)**self.gamma * logpt
-        if self.size_average: return loss.mean()
-        else: return loss.sum()
 
 
 class MultiBoxLoss(nn.Module):
@@ -79,15 +38,9 @@ class MultiBoxLoss(nn.Module):
         self.negpos_ratio = cfg.NEG_POS_RATIOS
         self.variance = cfg.VARIANCE
         self.dataset = dataset
-        if dataset == 'face':
-            self.threshold = cfg.FACE.OVERLAP_THRESH
-            self.match = match
-        elif dataset == 'hand':
-            self.threshold = cfg.HAND.OVERLAP_THRESH
-            self.match = match_ssd
-        else:
-            self.threshold = cfg.HEAD.OVERLAP_THRESH
-            self.match = match
+        
+        self.threshold = cfg.FACE.OVERLAP_THRESH
+        self.match = match
 
     def forward(self, predictions, targets):
         """Multibox Loss
@@ -154,121 +107,6 @@ class MultiBoxLoss(nn.Module):
                            ].view(-1, self.num_classes)
         targets_weighted = conf_t[(pos + neg).gt(0)]
         loss_c = F.cross_entropy(conf_p, targets_weighted, size_average=False)
-
-        # Sum of losses: L(x,c,l,g) = (Lconf(x, c) + αLloc(x,l,g)) / N
-        N = num_pos.data.sum() if num_pos.data.sum() > 0 else num
-        loss_l /= N
-        loss_c /= N
-        return loss_l, loss_c
-
-
-class MultiBoxLossFocal(nn.Module):
-    """SSD Weighted Loss Function
-    Compute Targets:
-        1) Produce Confidence Target Indices by matching  ground truth boxes
-           with (default) 'priorboxes' that have jaccard index > threshold parameter
-           (default threshold: 0.5).
-        2) Produce localization target by 'encoding' variance into offsets of ground
-           truth boxes and their matched  'priorboxes'.
-        3) Hard negative mining to filter the excessive number of negative examples
-           that comes with using a large number of default bounding boxes.
-           (default negative:positive ratio 3:1)
-    Objective Loss:
-        L(x,c,l,g) = (Lconf(x, c) + αLloc(x,l,g)) / N
-        Where, Lconf is the CrossEntropy Loss and Lloc is the SmoothL1 Loss
-        weighted by α which is set to 1 by cross val.
-        Args:
-            c: class confidences,
-            l: predicted boxes,
-            g: ground truth boxes
-            N: number of matched default boxes
-        See: https://arxiv.org/pdf/1512.02325.pdf for more details.
-    """
-
-    def __init__(self, cfg, dataset, use_gpu=True):
-        super(MultiBoxLossFocal, self).__init__()
-        self.use_gpu = use_gpu
-        self.num_classes = cfg.NUM_CLASSES
-        self.negpos_ratio = cfg.NEG_POS_RATIOS
-        self.variance = cfg.VARIANCE
-        self.dataset = dataset
-        if dataset == 'face':
-            self.threshold = cfg.FACE.OVERLAP_THRESH
-            self.match = match
-        elif dataset == 'hand':
-            self.threshold = cfg.HAND.OVERLAP_THRESH
-            self.match = match_ssd
-        else:
-            self.threshold = cfg.HEAD.OVERLAP_THRESH
-            self.match = match
-        self.focal_loss = FocalLoss(gamma=1, alpha=0.25, size_average=False)
-
-    def forward(self, predictions, targets):
-        """Multibox Loss
-        Args:
-            predictions (tuple): A tuple containing loc preds, conf preds,
-            and prior boxes from SSD net.
-                conf shape: torch.size(batch_size,num_priors,num_classes)
-                loc shape: torch.size(batch_size,num_priors,4)
-                priors shape: torch.size(num_priors,4)
-
-            targets (tensor): Ground truth boxes and labels for a batch,
-                shape: [batch_size,num_objs,5] (last idx is the label).
-        """
-        loc_data, conf_data, priors = predictions
-        num = loc_data.size(0)
-        priors = priors[:loc_data.size(1), :]
-        num_priors = (priors.size(0))
-        num_classes = self.num_classes
-
-        # match priors (default boxes) and ground truth boxes
-        loc_t = torch.Tensor(num, num_priors, 4)
-        conf_t = torch.LongTensor(num, num_priors)
-        for idx in range(num):
-            truths = targets[idx][:, :-1].data
-            labels = targets[idx][:, -1].data
-            defaults = priors.data
-            self.match(self.threshold, truths, defaults, self.variance, labels,
-                       loc_t, conf_t, idx)
-        if self.use_gpu:
-            loc_t = loc_t.cuda()
-            conf_t = conf_t.cuda()
-        # wrap targets
-        loc_t = Variable(loc_t, requires_grad=False)
-        conf_t = Variable(conf_t, requires_grad=False)
-
-        pos = conf_t > 0
-        num_pos = pos.sum(dim=1, keepdim=True)
-        # Localization Loss (Smooth L1)
-        # Shape: [batch,num_priors,4]
-        pos_idx = pos.unsqueeze(pos.dim()).expand_as(loc_data)
-        loc_p = loc_data[pos_idx].view(-1, 4)
-        loc_t = loc_t[pos_idx].view(-1, 4)
-        loss_l = F.smooth_l1_loss(loc_p, loc_t, size_average=False)
-        # print(loc_p)
-        # Compute max conf across batch for hard negative mining
-        
-        batch_conf = conf_data.view(-1, self.num_classes)
-        loss_c = log_sum_exp(batch_conf) - \
-            batch_conf.gather(1, conf_t.view(-1, 1))
-
-        # Hard Negative Mining
-        loss_c[pos.view(-1, 1)] = 0  # filter out pos boxes for now
-        loss_c = loss_c.view(num, -1)
-        _, loss_idx = loss_c.sort(1, descending=True)
-        _, idx_rank = loss_idx.sort(1)
-        num_pos = pos.long().sum(1, keepdim=True)
-        num_neg = torch.clamp(self.negpos_ratio *
-                              num_pos, max=pos.size(1) - 1)
-        neg = idx_rank < num_neg.expand_as(idx_rank)
-
-        # Confidence Loss Including Positive and Negative Examples
-        pos_idx = pos.unsqueeze(2).expand_as(conf_data)
-        neg_idx = neg.unsqueeze(2).expand_as(conf_data)
-        conf_p = conf_data[(pos_idx + neg_idx).gt(0)
-                           ].view(-1, self.num_classes)
-        targets_weighted = conf_t[(pos + neg).gt(0)]
-        loss_c = self.focal_loss(conf_p, targets_weighted)
 
         # Sum of losses: L(x,c,l,g) = (Lconf(x, c) + αLloc(x,l,g)) / N
         N = num_pos.data.sum() if num_pos.data.sum() > 0 else num
